move subtype of test methods to two diferent classes
