+++ src/tests/Blef.GameLogic.Tests/PokerHands/HighCardOnTableTester.cs
@@ -4,7 +4,7 @@ using Xunit;
 
 namespace Blef.GameLogic.Tests.PokerHands
 {
-    public class HighCardTests
+    public class HighCardOnTableTester
     {
         [Fact]
         public void Given_table_without_high_card_Then_result_is_False()
@@ -45,17 +45,6 @@ namespace Blef.GameLogic.Tests.PokerHands
             bool isOnTable = highCard.IsOnTable(cardsOnTable);
 
             Assert.True(isOnTable);
-        }
-
-        [Fact]
-        public void When_comparing_better_hand_to_worse_hand_then_result_is_bigger_than_zero()
-        {
-            HighCard betterHand = new HighCard(new Card(Rank.Jack, Suit.Diamonds));
-            HighCard worseHand = new HighCard(new Card(Rank.Nine, Suit.Diamonds));
-
-            int result = betterHand.CompareTo(worseHand);
-
-            Assert.True(result > 0);
-        }
+        } 
     }
 }
